Deployment/requirements.txt
streamlit==1.10.0
numpy==1.22.4
opencv-python==4.6.0.66
pillow==9.1.1
tensorflow==2.9.1